requirements.txt
wandb==0.15.12
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
torchviz==0.0.2
datasets
python-dotenv